Android GUI update
- Export from report activity
- Set HTML as default view
- Project update
