--- Source/GUI/Android/app/src/main/res/layout/activity_export.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--  Copyright (c) MediaArea.net SARL. All Rights Reserved.
-
-      Use of this source code is governed by a BSD-style license that can
-      be found in the License.html file in the root of the source tree.
--->
-
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/export_layout"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    tools:context=".ExportActivity">
-
-    <android.support.constraint.ConstraintLayout
-        android:id="@+id/spinner_layout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/export_view_text"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:text="@string/format_text"
-            android:textAlignment="center"
-            android:layout_marginLeft="@dimen/text_margin"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent" />
-
-        <Spinner
-            android:id="@+id/export_view_spinner"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintLeft_toRightOf="@+id/export_view_text" />
-
-    </android.support.constraint.ConstraintLayout>
-
-    <Button
-        android:id="@+id/close_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="@color/colorPrimary"
-        android:text="@string/close_text"
-        app:layout_constraintRight_toLeftOf="@+id/export_btn"
-        app:layout_constraintTop_toBottomOf="@+id/spinner_layout"
-        style="@style/Widget.AppCompat.Button.Borderless" />
-
-    <Button
-        android:id="@+id/export_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="@color/colorPrimary"
-        android:text="@string/export_text"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/spinner_layout"
-        style="@style/Widget.AppCompat.Button.Borderless" />
-
-</android.support.constraint.ConstraintLayout>
